all tests working! (bar seeded values that need updating)
