+++ mace/tools/arg_parser.py
@@ -3,12 +3,25 @@
 # Authors: Ilyes Batatia, Gregor Simm, David Kovacs
 # This program is distributed under the MIT License (see MIT.md)
 ###########################################################################################
+from __future__ import annotations
 
 import argparse
 import os
 from typing import Optional
 
 from .default_keys import DefaultKeys
+
+
+def dict_to_arg_list(args_dict: dict) -> list[str]:
+    """Converts a dictionary of arguments to a command-line style list."""
+    result = []
+    for key, value in args_dict.items():
+        if value is None:
+            continue
+
+        result.append(f"--{key}")
+        result.append(str(value))
+    return result
 
 
 def build_default_arg_parser() -> argparse.ArgumentParser:
